requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.0.26
langchain>=0.1.0
langchain-openai>=0.0.2
sqlalchemy>=2.0.0
pandas>=2.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.6.0
